refactoring, split message location
