+++ src/main/java/it/gov/daf/km4city/Main.java
@@ -1,7 +1,11 @@
 package it.gov.daf.km4city;
 
+import it.gov.daf.km4city.api.ApiLocation;
+import org.json.simple.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.List;
 
 public class Main {
 
@@ -11,7 +15,8 @@ public class Main {
 
         ApiLocation apiLocation = new ApiLocation();
         try {
-            apiLocation.getLocation(43.7741,11.2453,43.7768,11.2515);
+            List<JSONObject> result = apiLocation.getLocationRecords(43.7741,11.2453,43.7768,11.2515);
+            logger.info("result {}",result);
         } catch (Exception e) {
             logger.error("error",e);
         }
