+++ src/main/java/it/gov/daf/km4city/api/ApiEvent.java
@@ -1,4 +1,4 @@
-package it.gov.daf.km4city;
+package it.gov.daf.km4city.api;
 
 import java.io.IOException;
